+++ src/tiktok_uploader/proxy_auth_extension/manifest.json
